Refactor and Security
Refactor User and Customer classes.
Change Encryption class to handle hash and salt through file.

+++ lending_tree_application/src/main/java/org/lendingtree/project/CustomerDatabase.java
@@ -103,7 +103,7 @@ public class CustomerDatabase {
         insertCustomer.setInt(8, newCustomer.getPaySlip());
         insertCustomer.setInt(9, newCustomer.getTaxDetails());
         insertCustomer.setString(10, newCustomer.getIdentificationNumber());
-        insertCustomer.setFloat(11, newCustomer.getRating());
+        insertCustomer.setDouble(11, newCustomer.getRating());
         insertCustomer.setString(12, newCustomer.getPassword());
 
         insertCustomer.executeUpdate();
